+++ extras/tests/JsonArray/copyArray.cpp
@@ -13,10 +13,10 @@ TEST_CASE("copyArray()") {
     int source[] = {1, 2, 3};
 
     bool ok = copyArray(source, array);
-    REQUIRE(ok);
+    CHECK(ok);
 
-    serializeJson(array, json, sizeof(json));
-    REQUIRE(std::string("[1,2,3]") == json);
+    serializeJson(array, json);
+    CHECK(std::string("[1,2,3]") == json);
   }
 
   SECTION("std::string[] -> JsonArray") {
@@ -26,10 +26,10 @@ TEST_CASE("copyArray()") {
     std::string source[] = {"a", "b", "c"};
 
     bool ok = copyArray(source, array);
-    REQUIRE(ok);
+    CHECK(ok);
 
-    serializeJson(array, json, sizeof(json));
-    REQUIRE(std::string("[\"a\",\"b\",\"c\"]") == json);
+    serializeJson(array, json);
+    CHECK(std::string("[\"a\",\"b\",\"c\"]") == json);
   }
 
   SECTION("int[] -> JsonDocument") {
@@ -38,10 +38,22 @@ TEST_CASE("copyArray()") {
     int source[] = {1, 2, 3};
 
     bool ok = copyArray(source, doc);
-    REQUIRE(ok);
+    CHECK(ok);
 
-    serializeJson(doc, json, sizeof(json));
-    REQUIRE(std::string("[1,2,3]") == json);
+    serializeJson(doc, json);
+    CHECK(std::string("[1,2,3]") == json);
+  }
+
+  SECTION("int[] -> MemberProxy") {
+    DynamicJsonDocument doc(4096);
+    char json[32];
+    int source[] = {1, 2, 3};
+
+    bool ok = copyArray(source, doc["data"]);
+    CHECK(ok);
+
+    serializeJson(doc, json);
+    CHECK(std::string("{\"data\":[1,2,3]}") == json);
   }
 
   SECTION("int[] -> JsonArray, but not enough memory") {
@@ -54,8 +66,8 @@ TEST_CASE("copyArray()") {
     bool ok = copyArray(source, array);
     REQUIRE_FALSE(ok);
 
-    serializeJson(array, json, sizeof(json));
-    REQUIRE(std::string("[1,2]") == json);
+    serializeJson(array, json);
+    CHECK(std::string("[1,2]") == json);
   }
 
   SECTION("int[][] -> JsonArray") {
@@ -65,10 +77,22 @@ TEST_CASE("copyArray()") {
     int source[][3] = {{1, 2, 3}, {4, 5, 6}};
 
     bool ok = copyArray(source, array);
-    REQUIRE(ok);
+    CHECK(ok);
 
-    serializeJson(array, json, sizeof(json));
-    REQUIRE(std::string("[[1,2,3],[4,5,6]]") == json);
+    serializeJson(array, json);
+    CHECK(std::string("[[1,2,3],[4,5,6]]") == json);
+  }
+
+  SECTION("int[][] -> MemberProxy") {
+    DynamicJsonDocument doc(4096);
+    char json[32];
+    int source[][3] = {{1, 2, 3}, {4, 5, 6}};
+
+    bool ok = copyArray(source, doc["data"]);
+    CHECK(ok);
+
+    serializeJson(doc, json);
+    CHECK(std::string("{\"data\":[[1,2,3],[4,5,6]]}") == json);
   }
 
   SECTION("int[][] -> JsonDocument") {
@@ -77,10 +101,10 @@ TEST_CASE("copyArray()") {
     int source[][3] = {{1, 2, 3}, {4, 5, 6}};
 
     bool ok = copyArray(source, doc);
-    REQUIRE(ok);
+    CHECK(ok);
 
-    serializeJson(doc, json, sizeof(json));
-    REQUIRE(std::string("[[1,2,3],[4,5,6]]") == json);
+    serializeJson(doc, json);
+    CHECK(std::string("[[1,2,3],[4,5,6]]") == json);
   }
 
   SECTION("int[][] -> JsonArray, but not enough memory") {
@@ -97,53 +121,53 @@ TEST_CASE("copyArray()") {
     CAPTURE(doc.memoryUsage());
     CHECK_FALSE(ok);
 
-    serializeJson(array, json, sizeof(json));
-    REQUIRE(std::string("[[1,2,3],[4,5]]") == json);
+    serializeJson(array, json);
+    CHECK(std::string("[[1,2,3],[4,5]]") == json);
   }
 
   SECTION("JsonArray -> int[], with more space than needed") {
     DynamicJsonDocument doc(4096);
     char json[] = "[1,2,3]";
     DeserializationError err = deserializeJson(doc, json);
-    REQUIRE(err == DeserializationError::Ok);
+    CHECK(err == DeserializationError::Ok);
     JsonArray array = doc.as<JsonArray>();
 
     int destination[4] = {0};
     size_t result = copyArray(array, destination);
 
-    REQUIRE(3 == result);
-    REQUIRE(1 == destination[0]);
-    REQUIRE(2 == destination[1]);
-    REQUIRE(3 == destination[2]);
-    REQUIRE(0 == destination[3]);
+    CHECK(3 == result);
+    CHECK(1 == destination[0]);
+    CHECK(2 == destination[1]);
+    CHECK(3 == destination[2]);
+    CHECK(0 == destination[3]);
   }
 
   SECTION("JsonArray -> int[], without enough space") {
     DynamicJsonDocument doc(4096);
     char json[] = "[1,2,3]";
     DeserializationError err = deserializeJson(doc, json);
-    REQUIRE(err == DeserializationError::Ok);
+    CHECK(err == DeserializationError::Ok);
     JsonArray array = doc.as<JsonArray>();
 
     int destination[2] = {0};
     size_t result = copyArray(array, destination);
 
-    REQUIRE(2 == result);
-    REQUIRE(1 == destination[0]);
-    REQUIRE(2 == destination[1]);
+    CHECK(2 == result);
+    CHECK(1 == destination[0]);
+    CHECK(2 == destination[1]);
   }
 
   SECTION("JsonArray -> std::string[]") {
     DynamicJsonDocument doc(4096);
     char json[] = "[\"a\",\"b\",\"c\"]";
     DeserializationError err = deserializeJson(doc, json);
-    REQUIRE(err == DeserializationError::Ok);
+    CHECK(err == DeserializationError::Ok);
     JsonArray array = doc.as<JsonArray>();
 
     std::string destination[4];
     size_t result = copyArray(array, destination);
 
-    REQUIRE(3 == result);
+    CHECK(3 == result);
     CHECK("a" == destination[0]);
     CHECK("b" == destination[1]);
     CHECK("c" == destination[2]);
@@ -154,16 +178,48 @@ TEST_CASE("copyArray()") {
     DynamicJsonDocument doc(4096);
     char json[] = "[1,2,3]";
     DeserializationError err = deserializeJson(doc, json);
-    REQUIRE(err == DeserializationError::Ok);
+    CHECK(err == DeserializationError::Ok);
 
     int destination[4] = {0};
     size_t result = copyArray(doc, destination);
 
-    REQUIRE(3 == result);
-    REQUIRE(1 == destination[0]);
-    REQUIRE(2 == destination[1]);
-    REQUIRE(3 == destination[2]);
-    REQUIRE(0 == destination[3]);
+    CHECK(3 == result);
+    CHECK(1 == destination[0]);
+    CHECK(2 == destination[1]);
+    CHECK(3 == destination[2]);
+    CHECK(0 == destination[3]);
+  }
+
+  SECTION("MemberProxy -> int[]") {
+    DynamicJsonDocument doc(4096);
+    char json[] = "{\"data\":[1,2,3]}";
+    DeserializationError err = deserializeJson(doc, json);
+    CHECK(err == DeserializationError::Ok);
+
+    int destination[4] = {0};
+    size_t result = copyArray(doc["data"], destination);
+
+    CHECK(3 == result);
+    CHECK(1 == destination[0]);
+    CHECK(2 == destination[1]);
+    CHECK(3 == destination[2]);
+    CHECK(0 == destination[3]);
+  }
+
+  SECTION("ElementProxy -> int[]") {
+    DynamicJsonDocument doc(4096);
+    char json[] = "[[1,2,3]]";
+    DeserializationError err = deserializeJson(doc, json);
+    CHECK(err == DeserializationError::Ok);
+
+    int destination[4] = {0};
+    size_t result = copyArray(doc[0], destination);
+
+    CHECK(3 == result);
+    CHECK(1 == destination[0]);
+    CHECK(2 == destination[1]);
+    CHECK(3 == destination[2]);
+    CHECK(0 == destination[3]);
   }
 
   SECTION("JsonArray -> int[][]") {
@@ -171,18 +227,18 @@ TEST_CASE("copyArray()") {
     char json[] = "[[1,2],[3],[4]]";
 
     DeserializationError err = deserializeJson(doc, json);
-    REQUIRE(err == DeserializationError::Ok);
+    CHECK(err == DeserializationError::Ok);
     JsonArray array = doc.as<JsonArray>();
 
     int destination[3][2] = {{0}};
     copyArray(array, destination);
 
-    REQUIRE(1 == destination[0][0]);
-    REQUIRE(2 == destination[0][1]);
-    REQUIRE(3 == destination[1][0]);
-    REQUIRE(0 == destination[1][1]);
-    REQUIRE(4 == destination[2][0]);
-    REQUIRE(0 == destination[2][1]);
+    CHECK(1 == destination[0][0]);
+    CHECK(2 == destination[0][1]);
+    CHECK(3 == destination[1][0]);
+    CHECK(0 == destination[1][1]);
+    CHECK(4 == destination[2][0]);
+    CHECK(0 == destination[2][1]);
   }
 
   SECTION("JsonDocument -> int[][]") {
@@ -190,16 +246,34 @@ TEST_CASE("copyArray()") {
     char json[] = "[[1,2],[3],[4]]";
 
     DeserializationError err = deserializeJson(doc, json);
-    REQUIRE(err == DeserializationError::Ok);
+    CHECK(err == DeserializationError::Ok);
 
     int destination[3][2] = {{0}};
     copyArray(doc, destination);
 
-    REQUIRE(1 == destination[0][0]);
-    REQUIRE(2 == destination[0][1]);
-    REQUIRE(3 == destination[1][0]);
-    REQUIRE(0 == destination[1][1]);
-    REQUIRE(4 == destination[2][0]);
-    REQUIRE(0 == destination[2][1]);
+    CHECK(1 == destination[0][0]);
+    CHECK(2 == destination[0][1]);
+    CHECK(3 == destination[1][0]);
+    CHECK(0 == destination[1][1]);
+    CHECK(4 == destination[2][0]);
+    CHECK(0 == destination[2][1]);
+  }
+
+  SECTION("MemberProxy -> int[][]") {
+    DynamicJsonDocument doc(4096);
+    char json[] = "{\"data\":[[1,2],[3],[4]]}";
+
+    DeserializationError err = deserializeJson(doc, json);
+    CHECK(err == DeserializationError::Ok);
+
+    int destination[3][2] = {{0}};
+    copyArray(doc["data"], destination);
+
+    CHECK(1 == destination[0][0]);
+    CHECK(2 == destination[0][1]);
+    CHECK(3 == destination[1][0]);
+    CHECK(0 == destination[1][1]);
+    CHECK(4 == destination[2][0]);
+    CHECK(0 == destination[2][1]);
   }
 }